requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==10.4.0
python-decouple==3.8
requests==2.32.3
twilio==9.3.6
